Extract proof closeout runtime packets
